Done week 2 Sprint
I have successfully made a ground up tower with a very castle like features. I have made the map in which the enemies will spawn on.  I have coded a few things regrading the fps, even though it is not my part to do, I did it In order to learn a few things about FPS controlling systems and code :+1: